fix-double patching error
