+++ packages/instrumentation-mongodb/package.json
@@ -19,10 +19,10 @@
     "test": "npm run test-v6 && npm run test-unit",
     "test-unit": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/unit/*.test.ts",
     "test-v3": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-v3.test.ts",
-    "test-v4": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-v4.test.ts",
-    "test-v5": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-v5plus.test.ts",
-    "test-v6": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-v5plus.test.ts",
-    "test-v7": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-v5plus.test.ts",
+    "test-v4": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-metrics-stable-semconv-only.test.ts test/mongodb-v4.test.ts",
+    "test-v5": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-metrics-stable-semconv-only.test.ts test/mongodb-v5plus.test.ts",
+    "test-v6": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-metrics-stable-semconv-only.test.ts test/mongodb-v5plus.test.ts",
+    "test-v7": "nyc --no-clean mocha --require '@opentelemetry/contrib-test-utils' test/mongodb-metrics-v4plus.test.ts test/mongodb-metrics-stable-semconv-only.test.ts test/mongodb-v5plus.test.ts",
     "test:with-services-env": "cross-env NODE_OPTIONS='-r dotenv/config' DOTENV_CONFIG_PATH=../../test/test-services.env npm test",
     "test-all-versions": "tav",
     "test-all-versions:with-services-env": "cross-env NODE_OPTIONS='-r dotenv/config' DOTENV_CONFIG_PATH=../../test/test-services.env npm run test-all-versions",
